Remove compileAndRunToString from builtins-module-test
This test helper will break in an upcoming change when we
replace `FILE*` with Python file objects. This diff removes uses of
the helper, either by transforming the test into an equivalent test or by
noting that an equivalent test exists elsewhere and deleting the
`compileAndRunToString` test.

Based on Facebook D16572718