Add Latin-1 to the codec fast-path table
The `decode` and `encode` functions (in codecs and in builtins) first look through a table of popular encoding names mapped to their respective functions. If they can't find the names there, they take the slow path of relying on the `encodings` module.

All decoders/encoders in `_codecs.py` should add their names to this fast-path table, so this diff adds `latin-1` to it.

Based on Facebook D16586511