str.__mod__: Support flags, width and precision for strings
Available flags:

`'#'`: The value conversion will use the “alternate form” (where defined below).
`'0'`: The conversion will be zero padded for numeric values.
`'-'`: The converted value is left adjusted (overrides the `'0'` conversion if both are given).
`' '`: (a space) A blank should be left before a positive number (or empty string) produced by a signed conversion.
`'+'`: A sign character (`'+'` or `'-'`) will precede the conversion (overrides a “space” flag).

Minimum field width (optional). If specified as an `'*'` (asterisk), the actual width is read from the next element of the tuple in values, and the object to convert comes after the minimum field width and optional precision.

Precision (optional), given as a `'.'` (dot) followed by the precision. If specified as `'*'` (an asterisk), the actual precision is read from the next element of the tuple in values, and the value to convert comes after the precision.

Based on Facebook D16522103