Filter out placeholders when accessing module dict keys
Module dicts can contain placeholders that are not actually an active item, but just created for caching purposes.

This diff creates module key accessing APIs that filter out these when accessing module dict keys. Iterator is used to reduce memory burden when we know we don't need to return the whole set of keys to the user code.

Based on Facebook D16582814