Add float parsing and formatting code
This is a copy of pystrtod.c and dtoa.c from cpython. The cpython code
itself is a copy of the venerable code by David M. Gay with some
additions for formatting variants.

We want our own version of this code because:
- libc has no direct API to influence string formatting short of printf format strings.
- The results are guaranteed to be the same as cpython.

The code is heavily modified to fit our coding style but has no
behavioral changes and I don't plan to do any.

Based on Facebook D16574220