Fix assignment bug in PyMem_Resize
Currently, `PyMem_Resize` is not reassigning the returned pointer which leads to incorrect behavior. Instead, let's reuse the implementation of `PyMem_Resize` from CPython to guarantee correctness.

An example of the current incorrect behavior is shown in the following example:
```
void* p = PyMem_New(...)
PyMem_Resize(p, ...);
PyMem_Resize(p, ...);
```

The first call to `PyMem_Resize` calls `realloc` which frees `p` but it never assigns the result of the `realloc`. Therefore, on the second call to `PyMem_Resize`, `p` will be `realloc` again, resulting in a double free.

Based on Facebook D16527689