Freeze _str_mod.py
If the built binary is used outside of the normal build directory and it uses `str.__mod__` it fails with: `_str_mod` not found. This freezes `_str_mod.py` so that it's always available with the binary itself.

This fixes the issue with `pyro_perf` not showing results

Based on Facebook D16664612