Add a DCHECK_IS_ON() macro for conditionalizing debug-only code
The intention is to make `DCHECK` somewhat independent of `NDEBUG` which
is useful for creating builds with checks enabled and the optimizer
turned on.

This also allows us to create variable scopes that are only present when
`DCHECK` is enabled

```
  if (DCHECK_IS_ON()) {
    int value = ...
    DCHECK(value != ...);
  }
```

without resorting to using an `#ifdef NDEBUG`.

Based on Facebook D16768404