Add bytesIsValidUTF8() and bytesIsValidStr()
Add function that tests whether a given sequence of bytes is suitable for a pyro str. This is the case when it is legal UTF-8 except that we also have to allow codepoints in the surrogate range which are technically not UTF-8. Hence the separate `bytesIsValidUTF8()` and `bytesIsValidStr()` functions here.

This code is heavily inspired by cpythons `utf8_decode` in `cpython/Objects/stringlib/codecs.h`.

Based on Facebook D16725764