Slightly simplify idx variable usage in _str_mod.format
I found that the code becomes slightly nicer, when `idx` points to the character currently in `c` rather than the next character that will be returned by the iterator.

I used the same indexing scheme in the parsing in D16777688 and thought I should bring this algorithm in sync.

Based on Facebook D16806188