Implement codecs.utf_8_decode and make PyUnicode_DecodeUTF8 use it
This implements `codecs.utf_8_decode` similar to other decoders.

The error handling capabilities of `isValidUtf8Codepoint` are taken mostly from CPython, to make sure we throw the same errors as them.

The `DecodeLocale` test had to be changed because its functionality relied on either a naïve implementation of `DecodeUTF8`, or on a complete implementation that supports the `surrogatepass` error handler. Since the `surrogatepass` error handler isn't important as of yet, we can defer that for later and simply switch the test to create the string in managed code, which will always be valid.

Based on Facebook D15811479