Add _PyFloat_Pack and _PyFloat_Unpack functions
Summary: This adds all the `_Float_Pack` and `_Float_Unpack` functions. This assumes that the system is little endian so it removes the code that tries to identify the endianess. Also the code was adapted to be formatted according to our standard. Aside from that it is a copy from the cpython implementation.

Based on Facebook D16824803