Fix wrong return when calling __import__ with submodules
Apparently `builtins.__import__` has to return the toplevel module when importing a submodule, i.e. `__import__("a.b.c")` has to return a reference to the `a` module.

The caching logic we had in there did not correctly handle this. This just removes the caching logic and leaves the details to `importlib` (there are still caching effects as importlib will find the already loaded modules in `sys.modules` as well).

Based on Facebook D16897884