Fix zlibmodule.c bug
When D16695049 converted zlib to be PEP-384 compliant, it converted a call to the deprecated `_PyLong_FromNbInt` followed by `PyLong_AsSsize_t` to a single call to `PyNumber_AsSsizeT`, but it did not convert the arguments.

This diff gets rid of `long_obj` and properly converts the input.

Based on Facebook D16898957