Move Frame::kCodeUnitSize to bytecode.h
`kCodeUnitSize` describes the size of an opcode and argument in the bytecode which in python3 is constant size for all ops. Move it from the `Frame` class into `bytecode.h` because it wasn't really related to the frame.

While on it replace a couple hardcoded `2` values in `bytecode.cpp`.

Based on Facebook D16922945