Implement builtins.dir with no arg
This change implements part of `dir()` for "Without arguments, return the list of names in the current local scope".

This case of `dir()` delegates most of its logic to `locals()`, and to share the code this diff creates `sys._getframe_locals()` to be shared across `locals()`, `vars()`, and `dir()`.

Based on Facebook D16927844