[codemod][pyro] migrate python3 ThriftServer to SvcThriftServer
ThriftServer does not load ServiceFramework, unlike SvcThriftServer, and not having ServiceFramework loaded is detrimental to both visibility and correct operation of thrift-py3 servers at Facebook.

ServiceFramework sets the `uex` and `uexw` headers in responses, which are required by ServiceRouter to properly treat user exceptions as errors. This is relevant when this is used for selection, retries, and even logging in general (error=0 for user exceptions without these headers).

In addition to this, ServiceFramework provides a bunch of counters that are extremely useful for analysing services in production, and are also required for most common CSLB configs to function (see P82303693 for an exhaustive list of the counters).

Based on Facebook D17052359