Enable cache for tests by default
Currently cache needs to be enabled for tests since it's disabled by default.

This change enables it for all tests, and converts a few test cases to disable cache when it's needed.

`Runtime`' constructor still accepts `cache_enabled` as its argument in case it's needed for testing. It will be eventually removed once we don't see any needs of that as cache invalidation gets more mature.

Based on Facebook D16970339