Reduce usage of private: in objects.h
objects.h exposes interfaces for low-level raw object manipulation. By it's nature it is a private piece only to be used within the runtime. People who use it need to know what they are doing anyway, because of that there is little motivation to make some things even more `private:`.

This is motivated by me occasionally running into situations where an object is known to be `RawSmallStr`/`RawSmallBytes` but I have to assign to `RawStr`/`RawBytes` anyway because the methods of the small variants are private. This unnecessarily results in slightly less efficient code.

Based on Facebook D16900966