Use code point indexing for str subscripting
The `str.__getitem__` method defines its key arguments that are
not slices to be an index of a single code point.  However, our
implementation of this had been treating the key as the index of
a UTF-8 code unit.  That discrepancy allowed the user to
accidentally address part of a multi-byte character and the
resulting characters (a `str` instance of length 1) would contain
an invalid UTF-8 code unit.

This change converts the `int` key case of `str.__getitem__` to
addressing by code point and returning all of the code units underlying
a multi-byte character.

Based on Facebook D17060466