Implement locals() in non-function scopes
This change implements missing behaviors of `locals()` in module or class scopes. With this change, `locals()` returns a moduleproxy of the module in the calling function frame, and implicit globals if it's non-function or non-module scopes, which is represented by class scopes.

**Difference from CPython

This implementation is different from CPython in a way that `locals()` may return different instances in function and non-module scopes. This behavior is recorded in T53084807.

With this change, `import pickle` succeeds.

Based on Facebook D17085663