Make modular exponentiation more efficient
This change avoids taking the remainder of a potentially large value
of the base raised to a power.  Instead, the remainder is reduced as
it is being accumulated.

This avoids computing a modular exponention with a large modulus, such
as the following computation from `_pydecimal` feasible

`pow(10, 2305843009213693951 - 2, 2305843009213693951)`

Based on Facebook D17144417