Implement PyType_FromSpec subclassing
Modify `PyType_FromSpec` to allow inheritance from the base
types. This implements the equivalent of `inherit_special` and
`inherit_slot` in CPython.

The supports inheritance of: type, number, sequence, and
mapping slots. Non function slots are also inherited such as
flags and size.

Note that the inheritance of `subclass_dealloc` was added but
the implementation has been purposely left unimplemented as
that is outside the scope of this change.

Based on Facebook D16610333