Eagerly allocate a managed proxy for extension instances
Move the allocation of a managed proxy from `ApiHandle::asObject`
to `PyObject_Init` which is the common path for `PyObject_{GC_}New`,
`PyObject_{GC_}NewVar`, and `PyType_GenericAlloc`.

This guarantees that every native object now has a managed instance
that can be used as a proxy in to the native instance in managed code.
Therefore, `ApiHandle::fromPyObject` can now directly pull a managed
object out of the `reference_` field, even when it's a native instance.

Since the `reference_` field will not be used to store a heap instance,
native instances have been modified to allocate extra header space
to create a linked list through all extension instances. This linked list
will be used to keep track of the object lifecycle of native instances. This
means that all the allocation and deallocation C-APIs had to be updated
to correctly handle this linked list.

Finally, `trackObject` was renamed to `trackNativeGcObject` to reflect
what it really does.

Based on Facebook D17174216