Remove dictHasEmptyItem check
Drop the `dictHasEmptyItem()` check for a 2x speedup in debug builds.

The check turned what should be O(1) dictionary accesses into O(n) operations the more packed the dictionaries were. The worst thing that could happen if there is no empty item in the dictionary is an endless loop which should be easy enough to identify and debug even without the check.

Based on Facebook D17192038