Change PyObject::reference_ to a uintptr_t
`reference_` usually holds a `RawObject`, which isn't a C++ pointer;
it's an opaque, pointer-sized value. This diff changes it to `uintptr_t` rather
than `void*` and fixes any build errors that resulted. It's somewhat similar to
D9927060 in spirit, with a much, much smaller footprint.

Most changes are `s/nullptr/0/` or removing casts. I did have to add one cast
in `Runtime::trackNativeObject()`, but that function is going to be changed a
lot in an upcoming diff and the cast will go away.

Based on Facebook D16973270