Use correct key comparison for C-API dict implementation
The existing implementation of the C-API dict tried to avoid doing a comparison using = by looking up and caching the `__eq__` method on the key being looked up, which is incorrect. Also the dict implementation has a bug of updating the existing key with the searched key when an update happens.

This change fixes the two aforementioned problem as follows:
- The correct behavior is to use `existing_key == key_being_looked_up`.
- This change also uses `builtins.hash` to get the hash code from the key being looked up, not via its `__hash__` function.
- Introduced `Dict::Bucket::setValue` to update only the value of an updated item.

**Details**
CPython uses `PyObject_RichCompareBool` (https://hlrz.com/source/xref/cpython-3.6/Objects/object.c?r=896c6357#751) with the existing key as the left operand and the searched key as the right operand, in order to perform key equality checks. This change replicates its behavior in our managed dict implementation.

Based on Facebook D17105474