Improve invalid function return result checking
Extension functions are not allowed to return NULL without a
pending exception and they are not allowed to return non-NULL
with a pending exception.  This adds the checks in both
situations and expands the checks to the wrapper functions which
had not been thoroughly checking their return value.

Based on Facebook D17220504