Add IcIterator to encapsulate caching iteration logic for cache invalidation
This change adds `IcIterator` that iterates through all cache entries as retaining the metadata for the current cache entry to access its key, value, and bytecode information. This is expected to simplify the ongoing development of caching invalidation for binary ops (D16952211).

Note that this iterator is designed only for cache invalidation since cache lookups are very perf-sensitive.

Based on Facebook D17123177