Implement subtype dealloc
Implement the subtype deallocation function for C Extensions.

The subtype dealloc is in charge of calling `tp_finalize`,
`tp_del`, the base type `tp_dealloc`, and to correctly set the
reference count of the instance type.

This follows `subtype_dealloc` in `typeobject.c` with a couple of
differences:

**Avoid the trashcan mechanism**:
In CPython, a dealloc of a container instance can trigger a chain of
deallocs. This may result in calls to managed code, potentially causing a
stack overflow. The trashcan mechanism segments the deallocations, to avoid
this problem. However, any call to managed code from subtypeDealloc means
that the instance is managed and it's deallocation is controlled by the
tracing GC. Therefore, it avoids having a chain of deallocs that could
lead to a stack overflow.

**GC and non-GC flagged instances**:
There is no need to specially handle the clearing of slots, weak references
or the instance dictionary. Therefore, both cases can be handled the same.
Each of these cases are explained below.

**Weak references**:
Any weak reference callbacks should be handled by the garbage collector
through the native instance proxy.

**Clear slots**:
The garbage collector should cleanly deallocate all the members contained
by the instance. There is no need to manually clean the slots.

**Decref instance dictionary**:
If an instance dictionary is present, it will be cleaned up by the garbage
collector through the native instance proxy.

Based on Facebook D17159961