Fix argument processing parameter name checking
The previous code checked argument name identity equality rather than
string equality, which meant that distinct but equal strings would compare
false and mess up the argument processing. This diff now calls `__eq__`
if one of the left or right side of the comparison is not an exact `str`.

Based on Facebook D17254278