Use MutableTuple for list items
Use the internal-only MutableTuple for storing list items.

This requires changing Runtime::dict{Items|Keys|Values} to return a list directly.
The two callsites (of dictKeys) that did not return a list both iterated over the tuple.
One of them does not even need the tuple, so we can save the allocation. In the other, build the tuple in place.
These changes fix the return type of PyMapping_{Items|Keys|Values}, which is no longer allowed to return a Tuple as of 3.7.

Based on Facebook D17237264