Remove incorrect native instance setattr test
Summary: This test is broken and makes D17242343 fail. In Pyro, the try successfully executes without raising. In CPython it raises. Adding a proper fix requires having a more concrete design on what `__setattr__` means on a native instance and where that attribute should go. Since this test is not easily fixable before the NativeProxy lands, this just removes the test and adds a task to track the behavior. This should be put back once we are correctly able to deal with native dictionaries.

Based on Facebook D17312407