Fix Py_tp_members T_CHAR bug
Currently, the Py_tp_member T_CHAR is not currently working. The instance setter is expecting a integer value to update the native instance but instead it receives a str type. This fixes the bug by giving the integer representation of the char to the native object setter.

This was uncovered by a typo in a test that has now been fixed.

Based on Facebook D17278814