Add hash parameter to Runtime::dictIncludes
`Runtime::dictIncludes` was previously calling `Runtime::hash` to compute the hash value of the key. This does not work correctly for any object overriding `__hash__`. There was a `dictIncludesWithHash` alternative but it wasn't used consistently, and some callers incorrectly called `dictIncludes`.

This patch addresses this by:
- Adding a `key_hash` parameteter to `dictIncludes` to force people to compute a hash (typically via `Interpreter::hash` that calls `__hash__`).
- Drop the now unnecessary `dictIncludesWitHash`.
- Fix callers of `dictIncludes`.

Based on Facebook D17143171