Create a NativeProxy type
This introduces the `NativeProxy` which represents the layout
of the heap instances that wrap native objects.

The `NativeProxy` layout consists of three slots:
* native: Points to the native instance
* dict: The dictionary of the native instance, used for extension types that declare a dictoffset.
* link: Used to thread together all the native instances that are scheduled for destruction.

This modifies the extension types to use the NativeProxy
as the base layout to always guarantee a constant offset to
the native instance.

Based on Facebook D17242343