Unwrap ModuleProxy in PyEval_EvalCode
This change adds support for `PyEval_EvalCode` to accept `module_proxy` type objects as `globals`. Also it adds unwrapping logic for `locals` in case it's the same as `globals` which is the case when evaluating code in the module scope.

This change is needed for `D17375073` to remove usages of `PyModule_GetDict` in Pyro. `PyModule_GetDict` should not be used in Pyro due to its containment of `ValueCells` in it.

Note that this change replicates the already implemented logic in Pyro for `builtins.exec`.

Based on Facebook D17240703