Add hash parameter to Runtime::dictRemove
`Runtime::dictRemove` was previously calling `Runtime::hash` to compute the hash value of the key. This does not work correctly for any object overriding `__hash__`. There was a `dictRemoveWithHash` alternative but it wasn't used consistently, and some callers incorrectly called `dictIncludes`.

This patch addresses this by:
- Adding a `key_hash` parameteter to `dictRemove` to force people to compute a hash (typically via `Interpreter::hash` that calls `__hash__`).
- Drop the now unnecessary `dictRemoveWitHash`.
- Fix callers of `dictRemove`.

Based on Facebook D17143661