Add specialized accessors for string keys to dictionary APIs
It is very common to have string keys when accessing dictionaries (for example the identifiers in `code.co_names`).

This change:
- Add `dictAtByStr`, `dictAtPutByStr`, `dictRemoveByStr`, `dictIncludesByStr` and `dictAtPutInValueCellByStr` to `Runtime`.
- Add `dictAtById` and `dictAtPutById` convenience functions.
- Change callers to use the specialized functions where possible.
- Because the type had to be changed anyway, I also renamed some variables from `Object key` to `Str name` where it made sense.

The change is motivated by an upcoming change that moves (potentially raising) `__hash__` calls out of the runtime. This will become simpler when the many cases that only deal with strings are special cased. There may also be a tiny performance advantage because we can directly call into `strHash()` though I don't expect that to be significant.

Note that these functions only work correctly for `str` instances, they do NOT work for instances of `str` subclasses. This also the reason for choosing `const Str& name` parameters.

Based on Facebook D17231968