Stub out missing functionality from test.support
I've commented out the `import`s and usages of modules that we can temporarily get away with not having, and modified `sys.version` to be closer to what `platform` expects.

A lot of the functionality in `test.support` is only used by modules that we don't need for Django, so we can stub those functions and classes out for now to test the higher-priority modules.

Also contains a temproary work-around for our lack of compiler flag forwarding.

Based on Facebook D17072679