Integrate _PyArg_ParseStack to facilitate use of METH_FASTCALL
This adds an implementation for `_PyArg_ParseStack` which has been adapted from cpython.

In order to integrate this function, I ended up adding a work around to remove a global static variable in cpython which was used as a cache of the parsers. As a byproduct of this issue and our desire to not modify an incoming static variable to assign a PyObject*, we do not construct parser->kwtuple in ParserInit() and reuse that PyObject* as is done in cpython. Instead, we create a local kwtuple and use that instance instead. We believe that we should be able to do away with this local tuple altogether by doing a direct string comparison and that refactor work has been documented in this task T54190943.

Based on Facebook D17283455