Simplify the interface for invalidating attribute caches
With this change, invalidation of attributes requires the value cell of
the attribute being invalidated.  This allows the invalidating function
to avoid having to lookup the value cell which was already known to
the caller.  Furthermore, we can now push the descriptor check into the
invalidation function since we have the value cell containing the value.

Based on Facebook D17436037