Return NoneType for None.__class__
Pyro currently has a bug of returning the property object wrapping `object.__class__` as is due to the fact that `property.__get__` cannot distinguish `instance` bound to `None` object intentionally for object get attr from instance bound to `None` for type get attr.

This diff sidesteps the problem by directly adding a type attribute to `NoneType`.

**Before this change**

```
[djang@devvm3983.prn3 ~/fbsource/fbcode/pyro/build] ./python
>>> None.__class__
<property object at 0x7efe84fcdb89>
```

**After**
```
[djang@devvm3983.prn3 ~/fbsource/fbcode/pyro/build] ./python
>>> None.__class__
<class 'NoneType'>
```

**Side-effect**

This change creates an effect deviating from CPython that adds `__class__` to `NoneType.__dict__`. If this change is approved, this behavior will be added to T53084807.

Based on Facebook D17408676