Cleanup native memory through the GC
Temporary fix to enable `import hmac`.

Currently, the lifecycle of native instances from C Extensions
is not properly managed. That means that some objects can
outlive the runtime and won't call `tp_dealloc` on them.
Therefore, leaking memory.

This now extends the GC to walk through the list of tracked
native instances and determines if the native proxy should
be collected.

This is a temporary fix since there are still cases that are not
handled correctly. For instance, the resurrection of instances
or instances reaching a reference count of 0 from native code
that are still pointed by another object in the heap.

This also added a couple of different fixes to enable this change:
* Return the reference count of the native instance to 1 and
cleaned up tests that expected the old behavior.

* Remove `PrintTracebackPrintsInvalidFrame` because the
runtime destruction now crashes when cleaning up on a
corrupt frame.

* Fixed `CallDelSlotFromManagedCode` which was meant to
call `tp_del` rather than `tp_finalize`.

Based on Facebook D17436677