Refactor PyType_FromSpec to use the runtime's typeNew
Currently `PyType_FromSpec` is creating a new type from scratch and setting every field manually which have caused bugs in the past such as not inheriting the flags from the parents, etc.

This change unifies `PyType_FromSpec` with the runtime's `typeNew`. This leaves everything that's already handled by the runtime to `typeNew` and everything that is extension related (i.e: slots) to `PyType_FromSpec`.

This is the first step in enabling C to managed inheritance and managed to C inheritance.

Based on Facebook D17472722