Implement mappingproxy type
This change implements `class mappingproxy` that wraps `typeproxy` to make its type conform to CPython's. This is necessary to implement for the following reason:

* `PyDictProxy_*` C-APIs handle this type.
* Type reflection in modules infer `mappingproxy` type from `type(type_obj.__dict__)`.
* Such obtained type of `mappingproxy` is used to instantiate a read-only `dict` in user code.

This change also includes a new under function, `_type_proxy_check` to do shortcutting creation of `mappingproxy` for a type proxy object.

Based on Facebook D17433386