Require interned string argument in instance get/set/del APIs
The layout system works on interned strings, interning is not a cheap operation so we should avoid doing it multiple times on the same string. To this end this diff changes the `instanceGet`/`instanceSet` APIs to simply accept the caller to provide an interned string instead of unconditionally interning itself.This way we can avoid unnecessary interning in contexts where the caller already has an interned string around.

- Change instance get/set/del API to expect interned string arguments.
- I changed the parameters from `Object& name` to `Str& name_interned` to make it more likely to trigger compiler errors for callers that do not call `internStr` on their parameters.
- Avoids interning multiple times in functions calling `layoutFindAttribute()` which interns internally that would later just intern again.

Based on Facebook D17550702