Replace isType() with isInstanceOfType() where necessary
`RawObject::isType()` will only return `true` for types that are a direct instance of `type` and `false` for instances of subclasses of type. In most situations type subclass instances should be recognized as well and `Runtime::isInstanceOfType()` should be used instead.

I reviewed all uses of `isType()` and replaced the ones that needed it with `isInstanceOfType()`.

Based on Facebook D17557647