Support bytearray in PyObject_GetBuffer()
This adds support for bytearray objects to `PyObject_GetBuffer()`.

Supporting this is problematic because the returned buffer should contain a C pointer to the bytearray buffer. This is not compatible with our compacting garbage collector that may move the buffer and thereby invalidate a previously returned pointer. For now we work-around the problem by copying the contents into a new buffer on the un-managed heap. I filed T54579154 describing the shortcomings of this solution and potential fixes in the future.

Based on Facebook D17578660