Pull in CPython's encoding module
Even though we can use CPython's `encoding` module, some of their encodings rely on binding module functions as classmethods, which we don't suppoort yet. This pulls in the module along with the encodings that we can support, and a later diff will patch `utf-8` to allow us to properly use `codecs.getincrementaldecoder('utf-8')`.

This means that our `codecs` lookup functions will be unable to find the codecs for files that we don't pull in. While users can still do `"test".encode("encoding")` if we have the appropriate encoder, they won't be able to do `codecs.lookup("encoding")` unless we pull in the `encoding` file from CPython.

Based on Facebook D17602852