Allow managed types and extension types to subclass each other
This enables:
* Managed from a C Extension Type subclassing
* C Extension Type to Managed subclassing

This adds a check to the type creation function to see if any of the
type's bases has a NativeProxy layout. If it does, this type will now be
considered an Extension Type as well. Given that, the layout is
modified by prepending three empty slots to match the `NativeProxy`
layout.

Note that this does not allow C Extension subclassing from builtin
types as it would create a layout conflict.

Based on Facebook D17240522