Separate native and GC native instance queues
Currently, extension types with `Py_TPFLAGS_HAVE_GC` are completely ignored and treated like normal extension types. This means that instances whose types have with the GC flag that go through the default  `PyType_GenericAlloc` are added to the incorrect tracking queue.

This bug was exposed when an extension type with `Py_TPFLAGS_HAVE_GC` had a dealloc function which called `PyObject_GC_UnTrack` to remove the instance from the queue. This resulted in a leak since we were adding it to the incorrect queue.

This fixes the bug by correctly tracking instances with the `Py_TPFLAGS_HAVE_GC` flag as well as adding proper lifecycle management for these instances.

Based on Facebook D17537456