Raise ValueError when raw I/O detached
Previously, if we tried to access `buf.closed` on a buffered I/O stream with the raw stream detached, we would get an `AttributeError` attempting to access `None.closed`.

Modify all of the properties and methods that access attributes on the raw stream to first check if the stream has been detached.

Based on Facebook D17676008