Implement _PyNamespace_New
This change implements `_PyNamespace_New` to return an instance of a managed type, `namespace`.

The only requirement for `namespace` is that any instance variable added to an instance of `namespace` needs to be stored in its backing dictionary. This can be done by enabling the overflow bit in an instance for namespace objects in the future.

Based on Facebook D17631297