Add RawFunction::moduleObject to carry module object function is defined in
`RawFunction::module` carries the name of the module where a function is defined, not the module object itself. This causes the following issue:

- `function.__globals__` cannot find its defining module so it just leaks module dict which violates all our caching invariances.
- While `exec(code, global_dict, local_dict)` is being executed, the executed code cannot find the module the code is defined in since the module doesn't exist in `Runtime::modules_`. This breaks module dict lookup functions like `locals()`, `vars()`, and their friends.
- This forces us to look up the module object in `Runtime::modules_` whenever a module object is looked up

This change adds `RawFunction::moduleObject` field to carry the defining module object for a function alongside with interface changes to `Thread::exec`, `Runtime::newFunctionWithCode`.

With this change `RawFunction::globals` becomes redundant. Removing it will be done in a follow-up diff in case an additional pointer dereference caused by that is not going to cause perf regression.

Follow-up diffs will fix each of the problems above.

Based on Facebook D17632349