Implement _io.BufferedRandom
This implementation does not match `_pyio.py` because we want `isinstance(buf_random, _io.BufferedReader)` and `isinstance(buf_random, _io.BufferedWriter)` to be false as in CPython.

This requires copy-pasting the implementation of many of the methods from `BufferedReader` and `BufferedWriter`.

Also adds a built-in layout for `BufferedRandom` and a `_memoryview_check` helper.

Based on Facebook D17677434