Set implicitGlobals to None when it equals globals
Pyro delegates name resolution in `locals` to `globals` if `frame->function.globals() == frame->implicitGlobals()`, and this matters to many opcodes (`*_NAME`, `*_ANNOTATION`, ...).

This change moves that check from interpreter to `Thread::exec` to set `implicitGlobals` to `None` when it can be resolved in `globals` instead.

This change makes the logic's intention clear and removes references to `function.globals()`.

Based on Facebook D17753480