Improve dumpExtendedHeapObject output
Summary: We did output something like `heap object <type "PyCapsule">:` which was easy to confuse and read as dumping a type object while we would in fact dump an object with that type. Change the debug dumper to format it like `heap object with layout xxx (<type "PyCapsule">)` now.

Based on Facebook D17773791