Fix bugs with extension instances in the GC
Currently, the scavenger is being overly aggressive with extension instance deallocation. This was needed when some extensions were leaking reference counts. Now that it has been fixed, this updates the scavenger signal to now look exactly for a `refcnt = 1` or if the native proxy is not being forwarded.

By fixing this, the following bug was uncovered:

* **More Memory leaks**: Extension instances were being passed back to manage without stealing the reference. These instances would never have their reference count fall below `2`. To fix this, `ApiHandle::stealReference` had to be modified to actually decrease the reference. Fixing this now uncovered the following bug:

* **Use after free**: The current `isForwarding` check is currently being done on the `to_object` rather than a `from_object`. This meant that `isForwarding` always evaluated to `false`. To fix this, `scavengePointer` was moved until after the `isForwarding` check. This makes the check now happen on the `to_object`.

Based on Facebook D17773216