Rename caller parameters to frame
The interpreter code tends to use `Frame* caller` as a parameter. It seems to me this was meant to distinguish between the caller frame and the callee frame in the process of calling a function. However this name is used all over the place in a lot of contexts that are not callee related. I rather would like to think of these things as a frame pointer and hence rename them to `Frame* frame` here.

The only exception being: `doCallXXX()`, `pushFrame()`, `popFrame()`, `builtinTrampolineImpl()` which are indeed in the business of setting up or tearing down callframes, for those I made sure they use `callee_frame` and `caller_frame` variables, everyone else I renamed to just `frame`.

Based on Facebook D17690290