Accept objects with __getitem__ as iterables throughout the intepreter
Every object with a `__getitem__` can be used as an iterable in cpython. There are several places in the interpreter that only search for `__iter__` and then give up. This makes them all use the same code and accept the same things as iterables.

This also makes the opcodes more correct by first checking whether the resulting iterator has a `__next__` method, and raising an error if it doesn't. This means that several `tailcall` opcodes must be made recursive to account for the check.

Based on Facebook D17744006