Enable cache for all unittests
Some unittests are disabling cache to execute the given bytecode as is, without being rewritten by bytecode rewriting.

This change introduces `runCodeNoBytecodeRewriting` to perform exactly that as we're enabling cache for all unittests. With this change, there's no unittest that disables caching, and this will make our test closer to production.

Based on Facebook D17821772