Fix TrampolinesTest.InterpreterClosureUsesCellValue to exercise tested code path
Currently `TrampolinesTest.InterpreterClosureUsesCellValue` has a problem of not running the bytecode that it's trying to run: it creates a function `foo` and never overwrites its `function.rewrittenBytecode`. As a result, it was just running code that returns the first const (for `pass`).

This change fixes the problem by actually creating a function that satisfies the precondition of the test (to guarantee to make it have a cellvar) via interpreter.

Based on Facebook D17821562