Fix extension type integral member limits
Summary: Using `Runtime::newInt()` with the maximum `unsigned long` or `unsigned long long` value overflows (to -1). Making the member accessors print truncation warnings. This change switches all the limits of unsigned types to `Runtime::newIntFromUnsigned` (it's technically not necessary for the smaller types but looks more consistent that way).

Based on Facebook D17856515