Fix strSlice to handle unicode strings
Summary: The current implementation of strSlice uses `str.charLength()` instead of `str.codePointLength()` to compute the number of elements in the slice. This problem causes a memory issue by reading out of the given buffer. This change fixes the bug by correctly using codePointLength() for slice calculation.

Based on Facebook D17869604