Prepare code for pushFrame() triggering a GC
Summary: The call related code is carefully tuned to avoid handles where possible. This means we have to carefully consider where a garbage collection could happen and invalidate the raw pointers in the code. So far the code assumed that `pushFrame()` cannot trigger a garbage collection. I am planning to change this soon so I am modifying the code to no longer assume the raw function reference stay valid until after `pushFrame()`. This is solved mostly by rewriting the `processCellvarsAndFreevars()` interface to just grab the function from the stackframe instead of getting it passed in as a parameter.

Based on Facebook D17880041